+++ src/main/java/umc/plantory/domain/chat/repository/ChatJpaRepository.java
@@ -1,4 +1,4 @@
-package umc.plantory.domain.chat.adapter.out.repository;
+package umc.plantory.domain.chat.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.plantory.domain.chat.entity.Chat;
