+++ src/main/java/umc/plantory/domain/chat/service/ChatCommandUseCase.java
@@ -1,4 +1,4 @@
-package umc.plantory.domain.chat.port.in;
+package umc.plantory.domain.chat.service;
 
 public interface ChatCommandUseCase {
     String ask(String content, Long memberId);
