+++ src/main/java/umc/plantory/global/apiPayload/exception/ExceptionAdvice.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+import umc.plantory.domain.chat.exception.ChatApiException;
 import umc.plantory.global.apiPayload.ApiResponse;
 import umc.plantory.global.apiPayload.code.ErrorReasonDTO;
 import umc.plantory.global.apiPayload.code.status.ErrorStatus;
